chore: Rename getPremiumSubscriptionUrl to getPremiumSubscriptionInfo
