+++ src/app/(proper_react)/(redesign)/(authenticated)/user/(dashboard)/dashboard/page.tsx
@@ -25,7 +25,7 @@ import {
   isEligibleForFreeScan,
   isEligibleForPremium,
 } from "../../../../../../functions/server/onerep";
-import getPremiumSubscriptionUrl from "../../../../../../functions/server/getPremiumSubscriptionUrl";
+import getPremiumSubscriptionUrl from "../../../../../../functions/server/getPremiumSubscriptionInfo";
 import { refreshStoredScanResults } from "../../../../../../functions/server/refreshStoredScanResults";
 import { getEnabledFeatureFlags } from "../../../../../../../db/tables/featureFlags";
 import { parseIso8601Datetime } from "../../../../../../../utils/parse";
